refactor: extract cli runtime package
